rewrites DbpfFile write routine to take traversable entries
This allows for passing iterators of entries to the write method so that
it is not necessary to cache all entries in, say, a stream. This has a
much better memory footprint as only the current entry (the one that
gets written) needs to be held in memory, but **previous entries can be
garbage-collected** and successive entries need not be constructed until
needed (the latter was the case before, already).

This also gets rid of hideous bugfix involving views.